fixed some warnings and a bug with 'double down' where player still have the option to do so after first hit

+++ blackjack/Deck.m
@@ -47,7 +47,7 @@
     NSMutableArray *shuffledCards = [NSMutableArray arrayWithCapacity:cards.count];
     NSMutableArray *cardCopy = [NSMutableArray arrayWithArray:cards];
     
-    for (int i = cards.count; i > 1; i--) {
+    for (NSInteger i = cards.count; i > 1; i--) {
         int randomIndex = arc4random() % i;
         Card *card = [cardCopy objectAtIndex:randomIndex];
         [cardCopy removeObjectAtIndex:randomIndex];
